core,services: move binary log logic to services
- some fields previously exposed as public for testing are now
  package private
- remove tests from ManagedChannelImplTest and ServerImplTest, they
  are not really needed since BinaryLogProviderImplTest already do
  very similar things
- No more static instance of BinaryLog.Factory, it will be created by
  BinaryLogProviderImpl's constructor
