+++ core/src/test/java/io/grpc/internal/ManagedChannelImplTest.java
@@ -62,8 +62,6 @@ import io.grpc.ConnectivityState;
 import io.grpc.ConnectivityStateInfo;
 import io.grpc.Context;
 import io.grpc.EquivalentAddressGroup;
-import io.grpc.ForwardingClientCall.SimpleForwardingClientCall;
-import io.grpc.ForwardingClientCallListener.SimpleForwardingClientCallListener;
 import io.grpc.IntegerMarshaller;
 import io.grpc.LoadBalancer;
 import io.grpc.LoadBalancer.Helper;
@@ -77,13 +75,10 @@ import io.grpc.MethodDescriptor;
 import io.grpc.MethodDescriptor.MethodType;
 import io.grpc.NameResolver;
 import io.grpc.SecurityLevel;
-import io.grpc.ServerInterceptor;
 import io.grpc.Status;
 import io.grpc.StringMarshaller;
 import io.grpc.internal.Channelz.ChannelStats;
-import io.grpc.internal.NoopClientCall.NoopClientCallListener;
 import io.grpc.internal.TestUtils.MockClientTransportInfo;
-import io.grpc.internal.testing.SingleMessageProducer;
 import java.net.SocketAddress;
 import java.net.URI;
 import java.util.ArrayList;
@@ -96,7 +91,6 @@ import java.util.concurrent.Executor;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.atomic.AtomicLong;
-import javax.annotation.Nullable;
 import org.junit.After;
 import org.junit.Assume;
 import org.junit.Before;
@@ -197,7 +191,6 @@ public class ManagedChannelImplTest {
   private ObjectPool<Executor> oobExecutorPool;
   @Mock
   private CallCredentials creds;
-  private BinaryLogProvider binlogProvider = null;
   private BlockingQueue<MockClientTransportInfo> transports;
 
   private ArgumentCaptor<ClientStreamListener> streamListenerCaptor =
@@ -237,7 +230,6 @@ public class ManagedChannelImplTest {
         .userAgent(userAgent);
     builder.executorPool = executorPool;
     builder.idleTimeoutMillis = idleTimeoutMillis;
-    builder.binlogProvider = binlogProvider;
     builder.channelz = channelz;
     checkState(channel == null);
     channel = new ManagedChannelImpl(
@@ -2306,189 +2298,6 @@ public class ManagedChannelImplTest {
     channel.shutdownNow();
     assertEquals(SHUTDOWN, getStats(channel).state);
     assertEquals(SHUTDOWN, getStats(oobChannel).state);
-  }
-
-  @Test
-  public void binaryLogInterceptor_eventOrdering() throws Exception {
-    final List<ClientInterceptor> recvInitialMetadata = new ArrayList<ClientInterceptor>();
-    final List<ClientInterceptor> sendInitialMetadata = new ArrayList<ClientInterceptor>();
-    final List<ClientInterceptor> sendReq = new ArrayList<ClientInterceptor>();
-    final List<ClientInterceptor> recvResp = new ArrayList<ClientInterceptor>();
-    final List<ClientInterceptor> recvTrailingMetadata = new ArrayList<ClientInterceptor>();
-    final class TracingClientInterceptor implements ClientInterceptor {
-      @Override
-      public <ReqT, RespT> ClientCall<ReqT, RespT> interceptCall(
-          MethodDescriptor<ReqT, RespT> method,
-          CallOptions callOptions,
-          Channel next) {
-        final ClientInterceptor ref = this;
-        return new SimpleForwardingClientCall<ReqT, RespT>(next.newCall(method, callOptions)) {
-          @Override
-          public void start(Listener<RespT> responseListener, Metadata headers) {
-            sendInitialMetadata.add(ref);
-            ClientCall.Listener<RespT> wListener =
-                new SimpleForwardingClientCallListener<RespT>(responseListener) {
-                  @Override
-                  public void onMessage(RespT message) {
-                    recvResp.add(ref);
-                    super.onMessage(message);
-                  }
-
-                  @Override
-                  public void onHeaders(Metadata headers) {
-                    recvInitialMetadata.add(ref);
-                    super.onHeaders(headers);
-                  }
-
-                  @Override
-                  public void onClose(Status status, Metadata trailers) {
-                    recvTrailingMetadata.add(ref);
-                    super.onClose(status, trailers);
-                  }
-                };
-            super.start(wListener, headers);
-          }
-
-          @Override
-          public void sendMessage(ReqT message) {
-            sendReq.add(ref);
-            super.sendMessage(message);
-          }
-        };
-      }
-    }
-
-    TracingClientInterceptor userInterceptor = new TracingClientInterceptor();
-    final TracingClientInterceptor binlogInterceptor = new TracingClientInterceptor();
-    binlogProvider = new BinaryLogProvider() {
-      @Nullable
-      @Override
-      public ServerInterceptor getServerInterceptor(String fullMethodName) {
-        return null;
-      }
-
-      @Override
-      public ClientInterceptor getClientInterceptor(String fullMethodName) {
-        return binlogInterceptor;
-      }
-
-      @Override
-      protected int priority() {
-        return 0;
-      }
-
-      @Override
-      protected boolean isAvailable() {
-        return true;
-      }
-    };
-
-    // perform an RPC
-    Metadata headers = new Metadata();
-    ClientStream mockStream = mock(ClientStream.class);
-    createChannel(
-        new FakeNameResolverFactory.Builder(expectedUri).build(),
-        ImmutableList.<ClientInterceptor>of(userInterceptor));
-    CallOptions options =
-        CallOptions.DEFAULT.withExecutor(executor.getScheduledExecutorService());
-    ClientCall<String, Integer> call = channel.newCall(method, options);
-    call.start(mockCallListener, headers);
-
-    Subchannel subchannel = helper.createSubchannel(addressGroup, Attributes.EMPTY);
-    subchannel.requestConnection();
-    MockClientTransportInfo transportInfo = transports.poll();
-    ConnectionClientTransport mockTransport = transportInfo.transport;
-    ManagedClientTransport.Listener transportListener = transportInfo.listener;
-    // binlog modifies the MethodDescriptor, so we can not use the same() matcher
-    when(mockTransport.newStream(
-        any(MethodDescriptor.class),
-        same(headers),
-        any(CallOptions.class)))
-        .thenReturn(mockStream);
-    transportListener.transportReady();
-    when(mockPicker.pickSubchannel(any(PickSubchannelArgs.class)))
-        .thenReturn(PickResult.withSubchannel(subchannel));
-    helper.updateBalancingState(READY, mockPicker);
-
-    executor.runDueTasks();
-    verify(mockStream).start(streamListenerCaptor.capture());
-
-    ClientStreamListener streamListener = streamListenerCaptor.getValue();
-    streamListener.headersRead(new Metadata());
-    assertEquals(1, executor.runDueTasks());
-    String actualRequest = "hello world";
-    call.sendMessage(actualRequest);
-    streamListener.messagesAvailable(new SingleMessageProducer(method.streamResponse(1234)));
-    assertEquals(1, executor.runDueTasks());
-    call.halfClose();
-    streamListener.closed(Status.OK, new Metadata());
-    assertEquals(1, executor.runDueTasks());
-    // end: perform an RPC
-
-    assertThat(recvInitialMetadata).containsExactly(binlogInterceptor, userInterceptor).inOrder();
-    assertThat(sendInitialMetadata).containsExactly(userInterceptor, binlogInterceptor).inOrder();
-    assertThat(sendReq).containsExactly(userInterceptor, binlogInterceptor).inOrder();
-    assertThat(recvResp).containsExactly(binlogInterceptor, userInterceptor).inOrder();
-    assertThat(recvTrailingMetadata).containsExactly(binlogInterceptor, userInterceptor);
-  }
-
-  @Test
-  public void binaryLogInterceptor_intercept_reqResp() throws Exception {
-    final class TracingClientInterceptor implements ClientInterceptor {
-      private final List<MethodDescriptor<?, ?>> interceptedMethods =
-          new ArrayList<MethodDescriptor<?, ?>>();
-
-      @Override
-      public <ReqT, RespT> ClientCall<ReqT, RespT> interceptCall(
-          MethodDescriptor<ReqT, RespT> method, CallOptions callOptions, Channel next) {
-        interceptedMethods.add(method);
-        return next.newCall(method, callOptions);
-      }
-    }
-
-    TracingClientInterceptor userInterceptor = new TracingClientInterceptor();
-    binlogProvider = new BinaryLogProvider() {
-      @Nullable
-      @Override
-      public ServerInterceptor getServerInterceptor(String fullMethodName) {
-        return null;
-      }
-
-      @Override
-      public ClientInterceptor getClientInterceptor(String fullMethodName) {
-        return new TracingClientInterceptor();
-      }
-
-      @Override
-      protected int priority() {
-        return 0;
-      }
-
-      @Override
-      protected boolean isAvailable() {
-        return true;
-      }
-    };
-    createChannel(
-        new FakeNameResolverFactory.Builder(expectedUri).build(),
-        Collections.<ClientInterceptor>singletonList(userInterceptor));
-    ClientCall<String, Integer> call =
-        channel.newCall(method, CallOptions.DEFAULT.withDeadlineAfter(0, TimeUnit.NANOSECONDS));
-    ClientCall.Listener<Integer> listener = new NoopClientCallListener<Integer>();
-    call.start(listener, new Metadata());
-    assertEquals(1, executor.runDueTasks());
-
-    String actualRequest = "hello world";
-    call.sendMessage(actualRequest);
-
-    // The user supplied interceptor must still operate on the original message types
-    assertThat(userInterceptor.interceptedMethods).hasSize(1);
-    assertSame(
-        method.getRequestMarshaller(),
-        userInterceptor.interceptedMethods.get(0).getRequestMarshaller());
-    assertSame(
-        method.getResponseMarshaller(),
-        userInterceptor.interceptedMethods.get(0).getResponseMarshaller());
   }
 
   private static class FakeBackoffPolicyProvider implements BackoffPolicy.Provider {
