+++ services/src/main/java/io/grpc/services/internal/BinaryLogSink.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package io.grpc.services;
+package io.grpc.services.internal;
 
 import com.google.protobuf.MessageLite;
 import java.io.Closeable;
