+++ services/src/main/resources/META-INF/services/io.grpc.internal.BinaryLogProvider
@@ -1,0 +1,1 @@
+io.grpc.services.internal.BinaryLogProviderImpl
